+++ backend/pkg/api/handle_users.go
@@ -14,7 +14,6 @@ import (
 	"net/http"
 
 	"github.com/cloudhut/common/rest"
-	"github.com/go-chi/chi/v5"
 	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/api/admin"
 )
 
@@ -31,7 +30,7 @@ func (api *API) handleGetUsers() http.HandlerFunc {
 	}
 	return func(w http.ResponseWriter, r *http.Request) {
 		// 1. Check if logged-in user is allowed to list Kafka users
-		canList, restErr := api.Hooks.Console.CanListKafkaUsers(r.Context())
+		canList, restErr := api.Hooks.Authorization.CanListKafkaUsers(r.Context())
 		if restErr != nil {
 			rest.SendRESTError(w, r, api.Logger, restErr)
 			return
@@ -59,7 +58,7 @@ func (api *API) handleGetUsers() http.HandlerFunc {
 			}
 			filteredUsers := make([]string, 0, len(users))
 			for _, user := range users {
-				if api.Hooks.Console.IsProtectedKafkaUser(user) {
+				if api.Hooks.Authorization.IsProtectedKafkaUser(user) {
 					continue
 				}
 				filteredUsers = append(filteredUsers, user)
@@ -110,7 +109,7 @@ func (api *API) handleCreateUser() http.HandlerFunc {
 		}
 
 		// 2. Check if logged-in user is allowed to create Kafka users
-		canCreate, restErr := api.Hooks.Console.CanCreateKafkaUsers(r.Context())
+		canCreate, restErr := api.Hooks.Authorization.CanCreateKafkaUsers(r.Context())
 		if restErr != nil {
 			rest.SendRESTError(w, r, api.Logger, restErr)
 			return
@@ -127,7 +126,7 @@ func (api *API) handleCreateUser() http.HandlerFunc {
 		}
 
 		// 3. Check if targeted user is a protected user
-		if api.Hooks.Console.IsProtectedKafkaUser(req.Username) {
+		if api.Hooks.Authorization.IsProtectedKafkaUser(req.Username) {
 			restErr := &rest.Error{
 				Err:      fmt.Errorf("requester tried to create a protected Kafka user"),
 				Status:   http.StatusForbidden,
@@ -164,7 +163,7 @@ func (api *API) handleCreateUser() http.HandlerFunc {
 
 func (api *API) handleDeleteUser() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		principalID := chi.URLParam(r, "principalID")
+		principalID := rest.GetURLParam(r, "principalID")
 		if principalID == "" {
 			rest.SendRESTError(w, r, api.Logger, &rest.Error{
 				Err:     fmt.Errorf("user must be set"),
@@ -175,7 +174,7 @@ func (api *API) handleDeleteUser() http.HandlerFunc {
 		}
 
 		// 2. Check if logged-in user is allowed to delete Kafka users
-		canDelete, restErr := api.Hooks.Console.CanDeleteKafkaUsers(r.Context())
+		canDelete, restErr := api.Hooks.Authorization.CanDeleteKafkaUsers(r.Context())
 		if restErr != nil {
 			rest.SendRESTError(w, r, api.Logger, restErr)
 			return
@@ -192,7 +191,7 @@ func (api *API) handleDeleteUser() http.HandlerFunc {
 		}
 
 		// 3. Check if targeted user is a protected user
-		if api.Hooks.Console.IsProtectedKafkaUser(principalID) {
+		if api.Hooks.Authorization.IsProtectedKafkaUser(principalID) {
 			restErr := &rest.Error{
 				Err:      fmt.Errorf("requester tried to delete a protected Kafka user"),
 				Status:   http.StatusForbidden,